Fix: Notes section doesn't throw error on last note / npc deleted. Additionally switch button for notes section test moved to common button testing file.
