+++ src/components/__tests__/detailsSessionNotes.test.tsx
@@ -1,11 +1,10 @@
-import { it, vi, expect, describe, afterEach } from "vitest";
+import { it, expect, describe, afterEach } from "vitest";
 import { cleanup, screen } from "@testing-library/react";
 import {
   MOCK_DATA_COLLECTION,
   MOCK_FUNCTIONS,
 } from "../../utils/test-mock-data";
 import { renderWithRouter } from "../../utils/test-utils";
-import { DetailsNotesButton } from "../session/detailsSession/DetailsNotesButton";
 import { DetailsNotesList } from "../session/detailsSession/DetailsNotesList";
 import { DetailsNotesListElement } from "../session/detailsSession/DetailsNotesListElement";
 import { DetailsNotesSharedInformation } from "../session/detailsSession/DetailsNotesSharedInformation";
@@ -21,27 +20,7 @@ afterEach(() => {
 });
 
 describe("renders session details notes & npcs components", () => {
-  it("should render working note / npc section switching buttons", () => {
-    const buttonSpy = vi.spyOn(MOCK_FUNCTIONS, "buttonFunction");
-    const label = "Notes";
-
-    renderWithRouter(
-      <DetailsNotesButton
-        label={label}
-        changeDisplayTo={1}
-        selectedWhen={1}
-        display={1}
-        setDisplay={MOCK_FUNCTIONS.buttonFunction}
-      />
-    );
-
-    const switchButton = screen.getByText(label);
-    expect(switchButton).toBeInTheDocument();
-    switchButton.click();
-    expect(buttonSpy).toHaveBeenCalled();
-  });
-
-  it("should render note / npc section list", () => {
+  it("Renders note / npc section list", () => {
     renderWithRouter(
       <DetailsNotesList
         dataId={mockNote.id}
@@ -54,7 +33,7 @@ describe("renders session details notes & npcs components", () => {
     expect(listElements.length).toBe(MOCK_DATA_COLLECTION.length);
   });
 
-  it("should render note / npc section's list element", () => {
+  it("Renders note / npc section's list element", () => {
     renderWithRouter(
       <DetailsNotesListElement
         id={mockNote.id}
@@ -68,7 +47,7 @@ describe("renders session details notes & npcs components", () => {
     expect(listElement).toBeInTheDocument();
   });
 
-  it("should render note / npc information", () => {
+  it("Renders note / npc information", () => {
     const data = MOCK_DATA_COLLECTION[0];
     const { note } = SESSION_FIELDS_CATEGORIES;
